# Btrfs volumes are ones the kernel mounts
The driver loads the data relocation tree on mount and refuses a volume
without it — "failed to read root (objectid=...551607): -2" — while
btrfs check never asks for it. That is why an image could pass the
checker and still not mount, and why this went unnoticed.

Three things had to be true, each only visible once the one before it
was. The root tree needs a ROOT_ITEM for the tree; the tree itself must
not be empty, so it carries the root directory inode and the ".."
reference back to itself that every fs tree root has; and the
superblock's used-byte total counts tree blocks, so leaving the new one
out of that count reads as "super bytes used mismatches actual used".

The object id is -9, which is unsigned on disk and negative as a signed
long. Both places that order leaf items compared them signed, which put
the new root first where btrfs expects it last — harmless-looking until
a binary search over a leaf ordered that way stops finding the roots
after it, which is what "could not find extent tree" was: the in-place
adder rebuilding the root tree in the wrong order.

Mounted with the kernel driver, a volume this writes now lists all six
probe files and hands back every byte; btrfs check passes as before.

+++ FileSystems/FileSystem.Btrfs/BtrfsInPlaceAdder.cs
@@ -1022,9 +1022,15 @@ public static class BtrfsInPlaceAdder {
     _ = numBytes;
   }
 
+  /// <remarks>
+  /// Object ids are unsigned on disk, and the data relocation tree's is -9 read
+  /// as a signed long. Sorting signed puts it first, where btrfs expects it
+  /// last — and a binary search over a leaf ordered that way stops finding the
+  /// roots that follow it, which reads as "could not find extent tree".
+  /// </remarks>
   private static void SortItems(List<Item> items) {
     items.Sort((a, b) => {
-      var c = a.ObjectId.CompareTo(b.ObjectId);
+      var c = ((ulong)a.ObjectId).CompareTo((ulong)b.ObjectId);
       if (c != 0) return c;
       c = a.Type.CompareTo(b.Type);
       if (c != 0) return c;
